[PATCH] Issue #2461695 (comment 65) by FooZee
Patch: https://www.drupal.org/files/issues/paragraphs-recursive_cloning_localised_paragraphs_field_collections-2461695-65.patch

Okay, I recreated the patch with all the code style stuff and some extra syntactical enhancements :)
Is it important that paragraphs are to be localized? What if the parent is translatable and the paragraph is not? Perhaps the paragraph does not need to be translatable (= localized). Different paragraph entities does not need to be translatable, as far as I can see.
yes, if a paragraphs (entity revision) field is marked as translatable, this means that in different languages it will (be able to) reference different paragraph entities ... what that really means is: it doesn't really matter whether the paragraph itself is translatable or not, we'll always pre-populate the widget and make it ready to create a new paragraph entity that will be only referenced from this host translation (even if it has the exact same content). I prefer doing it like this because it's much easier to understand from an editor perspective (it's more confusing as a content editor to have some paragraphs that I can change per translation and some that would change everywhere if changed in one translation)
It's unusual to pass the entity (type?) manager to the method. Does it change from one call to another?
The entity_manager is being passed just because it was already instantiated in a pervious point in the code before actually calling this function, it doesn't change, it comes from the IDE refactoring, I've moved this to a property on the widget object to be shareable between methods.
Is Drupal's entity reference protected against circular referencing? Otherwise we may need to check for this. (follow-up issue)
I think it's not, although I believe it's not exactly possible from the admin UI, yet I don't think it's impossible to create a circular reference programmatically, which really makes this issue interesting to investigate.