[PATCH] Issue #2461695 (comment 10) by Tom Robert
Patch: https://www.drupal.org/files/issues/paragraphs_field-translation-2461695.patch

I worked out a solution for my use case. And the most simple solution is don't translate the paragraph entities.
Wat i want is that the field of the host handles the translations and not the paragraph entities. When translating the host entity the paragraph entities are duplicated with the values of the original language instead of translated.
 // If there is no paragraph entity in this language create one and duplicate the fields.
      $target_langcode = $this->getCurrentLangcode($form_state, $items);
      if ($paragraphs_entity->language()->getId() != $target_langcode) {
        $values = $paragraphs_entity->toArray();
        $entity_type = $entity_manager->getDefinition($target_type);
        $bundle_key = $entity_type->getKey('bundle');
        $new_values = array(
          $bundle_key => $paragraphs_entity->bundle(),
          'langcode' => $target_langcode,
        );

        foreach ($paragraphs_entity->getFieldDefinitions() as $key => $value) {
          if (get_class($value) == 'Drupal\field\Entity\FieldConfig') {
            $new_values[$key] = $values[$key][0];
          }
        }
        $paragraphs_entity = $entity_manager->getStorage($target_type)->create($new_values);
       }

vs
    // If target translation is not yet available, populate it with data from the original paragraph.
      $target_langcode = $this->getCurrentLangcode($form_state, $items);
      if ($paragraphs_entity->language()->getId() != $target_langcode && !$paragraphs_entity->hasTranslation($target_langcode)) {
        $paragraphs_entity->addTranslation($target_langcode, $paragraphs_entity->toArray());
      }

      // Initiate the paragraph with the correct translation.
      if ($paragraphs_entity->hasTranslation($this->getCurrentLangcode($form_state, $items))) {
        $paragraphs_entity = $paragraphs_entity->getTranslation($this->getCurrentLangcode($form_state, $items));
      } else {
        $paragraphs_entity = $paragraphs_entity->addTranslation($this->getCurrentLangcode($form_state, $items));
      }

In this case the the paragraphs items can co-exist without conflicts. I don't there should be an obvious link between the two items. It is provided by the host field. In this way we can add delete and modify paragraph entities without it having an effect on the translation. If the host translation is deleted the paragraph items are deleted as wel. This is not the case with the translations in the current version.
The patch attached will implement this method of translation. But it destroys the other method when you want the host field same for all languages and translate the paragraph fields instead. I think we need check the language settings for the host-field. If it is set to be translated. Create different entities. If not provide an option if the paragraph fields should be translated.