[PATCH] Issue #2461695 (comment 15) by krlucas
Patch: https://www.drupal.org/files/issues/paragraphs_multilingual-2461695-15.patch

Attached is a re-roll of the patch from #11 which fixes the issue described in #13. It also copies all values for multi-valued nested paragraph fields, instead of just the first item.
I agree that the ideal solution to this translation strategy ("translate the primary host ERR field; never translate fields on paragraph entities") is to extend the InlineParagraphsWidget...once the work in #268239 is stable and committed.