[PATCH] Issue #2461695 (comment 34) by mikemiles86
Patch: https://www.drupal.org/files/issues/paragraphs-meta_support-2461695-34.patch

@miro_dietiker
I agree with you that what is being discussed is two different things, translation and localization. To your point we should be careful with how we mix those two.
Like others in this thread I have run into both of these situations with clients. Some would like purely translated content (same content different language) and others want localized content (different content for different languages).  Each has a content structure that would greatly benefit from using paragraphs.
Thanks in part to the way you have structured the module, I think the Paragraphs logic could be augmented to support localization in addition to translation with little need of altering the UI or worry about confusing the user. The following is my theory (as well as my patch) for how to simplistically support both use cases.
Scenario 1: Translated Paragraphs ( currently implemented)
In this Scenario the user does not translated the host (entity_reference) field, but instead translates the fields in the Paragraph Entity.  For the host entity this is a 1:1 translation. When the user creates a translation of the Entity, the paragraphs are translated as well. The user cannot alter the paragraphs in the translated Entity without altering them for all Entity translations (as in adding or reorganizing the paragraphs).
Scenario 2: Localized Paragraphs
In this Scenario the user does not translate the fields in the Paragraph Entity, instead they translate the host entity field (entity_reference field). When the user creates/edits a translation of an Entity, each paragraph is duplicated as a new paragraph entity in the translated language.  The user then has the ability to change the value in the entity Reference field (add, remove, edit, re-order the paragraphs) without effecting the other translations. Allowing them to 'localize' their paragraphs.
----
Setting up Paragraphs to work in this method requires no need for additional UI work as it allows paragraphs to follow the expected workflow when marking a field as translatable. If a field is marked as translatable you expect to be able to change the field value per translation. If a field is not translatable you expect it to contain the same value for all translations.
My attached patch add this intrinsic functionality. When translating a paragraph entity, then host field is checked. If the host field is marked as translatable the logic then assumes the user is looking to localize the paragraphs and creates new paragraph entities in that language and assigns their value to the field. When editing a translated entity the logic also checks if the host field is translatable, and if so (since it is assume the content is 'localized') enables access to manipulate and add more paragraphs for that translation.
This patch was very much influenced by the patch provided by @krlucas in comment #15, is almost basically a re-roll with some additional logic added.
Hope this adds positively to this feature and discussion.