[PATCH] Issue #2461695 (comment 59) by FooZee
Patch: https://www.drupal.org/files/issues/recursive_cloning_localised_paragraphs_field_collections-2461695-59.patch

This patch is based on #44 with the addition that it goes into the paragraphs being cloned and clones any paragraphs or field collections in them. I don't really like that we had to hard-code the entities we support for cloning although the code doing the cloning itself is pretty generic (but I had to do so, because it's very often not very desirable to clone other entity types referenced in the paragraph --e.g. users, taxonomy terms and probably nodes-- ) ...
In fact, whether an entity should be cloned or just referenced (when we are speaking about asymmetric translation) largely depends on whether the widget shown on that form allows the entity to be edited (but that's not a very practical condition, because a site builder can always change the widget of such an entity back and forth, which would leave the DB totally inconsistent) thoughts specifically about this are highly appreciated :)