[PATCH] Issue #2461695 (comment 121) by Rene Bakx
Patch: https://www.drupal.org/files/issues/meta_support-2461695-against-dev-120.patch

I was researching the localized options for Paragraphs, mostly for the reasons like explained above. Our current project requires that the editor should be able to add/remove a paragraph per translated node.
Tried to roll the #106 patch against the current dev version, but that failed so I rolled the failed parts of the patch by hand. Hope I got them all. (long live pull requests instead of patches)
See
Anyhow, attached you that patch, hopefully it works. And for what it is worth, hopes this will be the defacto behaviour for Paragraphs or else at least will continues to live as a separate community project as a separate widget because most of the 'magic' happens in there. The form overrides could be de-overridden in a .module.