[PATCH] Issue #2461695 (comment 128) by hoebekewim
Patch: https://www.drupal.org/files/issues/meta_support-2461695-against-dev-128.patch

I've re-rolled the patch id 120 from comment #121 since it failed on the DEV version. Further I've implemented the following fixes:
In the scenario where you create a node in language 1 and then translate it, it is not possible to add more paragraphs into the translation. This is because the add buttons are not shown when translating a node. I've created an exception to check if the field is set as "translatable" before showing the add buttons. This way the add buttons won't show up on a paragraphs field that is not set as translatable.
The comment from #127 referencing to #105 has also been applied. I don't know what this does, can someone check if this does not break anything in nested paragraphs ?