[PATCH] Issue #2461695 (comment 184) by efpapado
Patch: https://www.drupal.org/files/issues/meta_support-2461695-184.patch

I submit a new patch that:
Adds a test for nested paragraphs (one level of nesting)
Adds a test for deleting a translated node
Removes some trailing spaces that were introduced by my previous patch (this can be seen at the interdiff)
Please note that an important change was implemented, in order to test for nested paragraphs: The paragraph entity, and it's paragraph reference field had to be set as translatable. I left the previous configuration (that was setting them as non-translatable) commented on purpose.